Changes to routing with url endpoints.
