+++ src/constants/routes.js
@@ -2,3 +2,7 @@
 export const SIGN_UP = '/signup';
 export const SIGN_IN = '/login';
 export const HOME = '/';
+export const EVENTS = '/events';
+export const POINTS = '/points';
+export const RESUME = '/resume';
+export const INDUCTEES = '/inductees';
